Loads (if possible) next level when current level is beaten

+++ src/arkanoid/logic/entity/ball/ball.cpp
@@ -31,7 +31,7 @@ namespace arkanoid {
 		}
 	}
 
-	void Ball::setInvisible(double period) {
+	void Ball::setInvisible(int period) {
 		invisDuration = period;
 	}
